requirements.txt
peewee==3.17.0
python-dotenv==1.0.0
rich==13.6.0
urllib3==2.0.6
